+++ Src/ModuleAssemblyInitializer.cs
@@ -3,11 +3,11 @@
 // http://www.apache.org/licenses/LICENSE-2.0
 
 using MongoDB.Bson;
+using MongoDB.Bson.IO;
 using MongoDB.Bson.Serialization;
+using MongoDB.Bson.Serialization.Serializers;
 using System;
-using System.IO;
 using System.Management.Automation;
-using System.Reflection;
 
 namespace Mdbc;
 
@@ -20,40 +20,17 @@ public class ModuleAssemblyInitializer : IModuleAssemblyInitializer
 
 	static ModuleAssemblyInitializer()
 	{
-		AppDomain.CurrentDomain.AssemblyResolve += AssemblyResolve;
+		// Changed from Shell to RelaxedExtendedJson in 3.0, not useful for interactive `ToString()`, etc.
+		if (!JsonWriterSettings.Defaults.IsFrozen)
+			JsonWriterSettings.Defaults.OutputMode = JsonOutputMode.Shell;
 
-		BsonSerializer.RegisterSerializer(typeof(Collection), new CollectionSerializer());
 		BsonSerializer.RegisterSerializer(typeof(Dictionary), new DictionarySerializer());
+		BsonSerializer.RegisterSerializer(typeof(Collection), new CollectionSerializer());
+		BsonSerializer.RegisterSerializer(typeof(Guid), new GuidSerializer(Api.GuidRepresentation));
 		BsonSerializer.RegisterSerializer(typeof(PSObject), new PSObjectSerializer());
+		BsonSerializer.RegisterSerializer(typeof(object), new ObjectSerializer(ObjectSerializer.Instance.DiscriminatorConvention, Api.GuidRepresentation));
 
+		BsonTypeMapper.RegisterCustomTypeMapper(typeof(Guid), new GuidTypeMapper());
 		BsonTypeMapper.RegisterCustomTypeMapper(typeof(PSObject), new PSObjectTypeMapper());
-
-		var strGuidRepresentation = Environment.GetEnvironmentVariable("Mdbc_GuidRepresentation");
-#pragma warning disable 618 // obsolete BsonDefaults.GuidRepresentation
-		if (strGuidRepresentation == null)
-		{
-			BsonDefaults.GuidRepresentation = GuidRepresentation.Standard;
-		}
-		else
-		{
-			if (Enum.TryParse(strGuidRepresentation, out GuidRepresentation valGuidRepresentation))
-				BsonDefaults.GuidRepresentation = valGuidRepresentation;
-			else
-				throw new InvalidOperationException($"Invalid environment variable Mdbc_GuidRepresentation = {strGuidRepresentation}");
-		}
-#pragma warning restore 618
-	}
-
-	// Workaround for Desktop
-	static Assembly AssemblyResolve(object sender, ResolveEventArgs args)
-	{
-		if (args.Name.StartsWith("System.Runtime.CompilerServices.Unsafe"))
-		{
-			var root = Path.GetDirectoryName(typeof(ModuleAssemblyInitializer).Assembly.Location);
-			var path = Path.Combine(root, "System.Runtime.CompilerServices.Unsafe.dll");
-			var assembly = Assembly.LoadFrom(path);
-			return assembly;
-		}
-		return null;
 	}
 }